% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{gen.clusters}
\alias{gen.clusters}
\title{Generate clusters}
\usage{
gen.clusters(pars, K, ICC, njbar)
}
\arguments{
\item{pars}{are the sample mean and sd's for the underlying regression
parameters. pars cotains means, sds, and names}

\item{K}{is the total clusters in the sample}

\item{ICC}{is the ICC of the running variable}

\item{njbar}{is the mean total observations in each cluster}
}
\description{
Generate clusters
}
