% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{gen.data}
\alias{gen.data}
\title{Generate set of clusters with RDD in each clustrer.}
\usage{
gen.data(TC = 150, nmean = 130, parameters.reg, obs.se, Rmean)
}
\arguments{
\item{TC}{total clusters (TC) to 150}

\item{nmean}{average observations per cluster (nmean) to 130}
}
\description{
Generate set of clusters with RDD in each clustrer.
}
