% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_code.R
\name{gen.obs}
\alias{gen.obs}
\title{generate a sample of students using the school parameters}
\usage{
gen.obs(a0, b0, a1, b1, rj, ICC, Nj, cl.id, TC, nmean, se, rmean)
}
\arguments{
\item{se}{is the student level standard error}

\item{rmean}{is the grand mean of the running variable}

\item{pars}{are the parameters for each school from gen.schools}
}
\description{
assume a cut point of 0, treatment if below cut
}
