% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_library.R
\name{Q.Stat.Inv}
\alias{Q.Stat.Inv}
\title{Q-Stat Inverstion Function}
\usage{
Q.Stat.Inv(bj, vj)
}
\arguments{
\item{bj}{List of estimated impacts}

\item{vj}{List of associated standard errors.}
}
\description{
Q-Stat Inverstion Function
}
