% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_library.R
\name{reg.FIRC}
\alias{reg.FIRC}
\title{Conduct FIRC analysis of multi RDD}
\usage{
reg.FIRC(
  df,
  cutpoint = 0,
  less.than = FALSE,
  clusterid,
  runvar,
  outcomevar,
  model
)
}
\arguments{
\item{df}{The dataset to analyze.}
}
\description{
Conduct FIRC analysis of multi RDD
}
