% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_library.R
\name{reg.LLR}
\alias{reg.LLR}
\title{LLR}
\usage{
reg.LLR(df, clusterid, cutpoint = 0, runvar, outcomevar, less.than = FALSE)
}
\arguments{
\item{df}{The dataset to analyze.}
}
\description{
Before running this method, your data needs to be subsetted to just in
bandwidth obs and small clusters dropped before.
}
