% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_library.R
\name{reg.Meta}
\alias{reg.Meta}
\title{Conduct meta-analysis of multi RDD}
\usage{
reg.Meta(df, cutpoint = 0, less.than = FALSE, clusterid, runvar, outcomevar)
}
\arguments{
\item{df}{The dataset to analyze.}
}
\description{
drop small clusters and subset to only data in bw before running data
}
